fix(db_commit): deleted extra commit

+++ backend/tools/for_db/work_with_slots.py
@@ -30,8 +30,9 @@ def get_id_slice_of_slot(start, end, admin_id):
             raise DbSlotException('Error. Slot already booked')
         start_db = slot_from_db.start_interval
         end_db = slot_from_db.end_interval
+        if start_db == start and end_db == end:
+            return slot_from_db.id
         models.db.session.delete(slot_from_db)
-        models.db.session.commit()
         if start_db < start:
             slot1 = Slots(start_interval=start_db, end_interval=start)
             models.db.session.add(slot1)
